Methods

Methods

This section describes systems methods implemented in PolynomialZonotopes.jl.

Methods

Properties

Operations

Properties

PolynomialZonotopes.dim — Method.

dim(pz::PolynomialZonotope)::Int

Return the ambient dimension of a polynomial zonotope.

Input

pz – polynomial zonotope

Output

An integer representing the ambient dimension of the polynomial zonotope.

source

PolynomialZonotopes.σ — Method.

σ(d::AbstractVector{N}, pz::PolynomialZonotope{N})::AbstractVector{N} where {N}

Return the support vector of a polynomial zonotope along direction d.

Input

d – direction

pz – polynomial zonotope

Output

Vector representing the support vector.

source

PolynomialZonotopes.polynomial_order — Method.

polynomial_order(pz::PolynomialZonotope)::Int

Polynomial order of a polynomial zonotope.

Input

pz – polynomial zonotope

Output

The polynomial order, defined as the maximal power of the scale factors $β_i$. Usually denoted $η$.

source

PolynomialZonotopes.order — Method.

order(pz::PolynomialZonotope)::Rational{Int}

Order of a polynomial zonotope.

Input

pz – polynomial zonotope

Output

The order, defined as the number of generators divided by the ambient dimension.

source

Operations

PolynomialZonotopes.linear_map — Method.

linear_map(M::Matrix, pz::PolynomialZonotope)

Return the linear map of a polynomial zonotope.

Input

M – matrix

pz – polynomial zonotope

Output

Polynomial zonotope such that its starting point and generators are those of pz multiplied by the matrix M.

source

PolynomialZonotopes.scale — Method.

scale(α::Number, pz::PolynomialZonotope)

Return a polynomial zonotope modified by a scale factor.

Input

α – polynomial zonotope

pz – polynomial zonotope

Output

Polynomial zonotope such that its center and generators are multiples of those of pz by a factor $α$.

source

PolynomialZonotopes.minkowski_sum — Method.

minkowski_sum(pz::PolynomialZonotope, z::Zonotope)

Return the Minkowski sum of a polynomial zonotope and a usual zonotope.

Input

pz – polynomial zonotope

z – usual zonotope

Output

Polynomial zonotope.

source